requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
pydantic[email]==2.4.2
bcrypt==4.0.1
openpyxl==3.1.2
reportlab==4.0.4
websockets==12.0
certifi==2023.11.17